% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2a.R
\name{a2a_stop}
\alias{a2a_stop}
\title{Stop an A2A server}
\usage{
a2a_stop(server)
}
\arguments{
\item{server}{An A2A server handle from \code{\link[=a2a_server]{a2a_server()}}}
}
\value{
\code{NULL}, invisibly
}
\description{
Stop an A2A server
}
