% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add_conditional_edge}
\alias{add_conditional_edge}
\title{Add a conditional edge to the graph}
\usage{
add_conditional_edge(graph, from, route)
}
\arguments{
\item{graph}{A graph object}

\item{from}{Source node ID}

\item{route}{A routing configuration (from route_on())}
}
\value{
The modified graph object
}
\description{
Add a conditional edge to the graph
}
