% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add_edge}
\alias{add_edge}
\title{Add an edge to the graph}
\usage{
add_edge(graph, from, to)
}
\arguments{
\item{graph}{A graph object}

\item{from}{Source node ID}

\item{to}{Target node ID (or "\strong{end}")}
}
\value{
The modified graph object
}
\description{
Add an edge to the graph
}
