% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{add_node}
\alias{add_node}
\title{Add a node to the graph}
\usage{
add_node(graph, id, node)
}
\arguments{
\item{graph}{A graph object}

\item{id}{Node ID}

\item{node}{A node configuration (from llm_node(), tool_node(), etc.)}
}
\value{
The modified graph object
}
\description{
Add a node to the graph
}
