% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{append_messages}
\alias{append_messages}
\title{Reducer: append messages}
\usage{
append_messages(existing, new)
}
\arguments{
\item{existing}{Existing messages}

\item{new}{New messages to append}
}
\value{
Combined messages
}
\description{
Reducer: append messages
}
