% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{assistant_msg}
\alias{assistant_msg}
\title{Create an assistant message}
\usage{
assistant_msg(content)
}
\arguments{
\item{content}{Message content}
}
\value{
A message list
}
\description{
Create an assistant message
}
