% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_status}
\alias{batch_status}
\title{Report the status of a batch job}
\usage{
batch_status(job)
}
\arguments{
\item{job}{A job from \code{\link[=batch_submit]{batch_submit()}}}
}
\value{
"running", "complete", or "error"
}
\description{
Report the status of a batch job
}
