% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{call_plugin}
\alias{call_plugin}
\title{Invoke a registered plugin factory}
\usage{
call_plugin(kind, name, ...)
}
\arguments{
\item{kind}{Plugin kind}

\item{name}{Plugin name}

\item{...}{Arguments passed to the factory}
}
\value{
The factory's return value
}
\description{
Invoke a registered plugin factory
}
