% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{content_parts}
\alias{content_parts}
\title{Combine content parts into a multimodal message body}
\usage{
content_parts(...)
}
\arguments{
\item{...}{One or more content parts (from \code{text_part()} / \code{image_part()})}
}
\value{
A list of content parts
}
\description{
Combine content parts into a multimodal message body
}
