% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{embed_batch}
\alias{embed_batch}
\title{Embed multiple texts in a single request}
\usage{
embed_batch(texts, provider)
}
\arguments{
\item{texts}{Character vector of texts to embed}

\item{provider}{A provider configuration}
}
\value{
A list of numeric vectors
}
\description{
Embed multiple texts in a single request
}
