% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_dataset}
\alias{feedback_dataset}
\title{Return the collected feedback as a data.frame}
\usage{
feedback_dataset()
}
\value{
A data.frame with columns \code{seq}, \code{ts_ms}, \code{input}, \code{output}, \code{rating},
\code{correction}, and \code{comment}
}
\description{
Return the collected feedback as a data.frame
}
