% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback_stats}
\alias{feedback_stats}
\title{Summarize feedback counts}
\usage{
feedback_stats()
}
\value{
A data.frame with one row: \code{good}, \code{bad}, \code{total}
}
\description{
Summarize feedback counts
}
