% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{fine_tune_cancel}
\alias{fine_tune_cancel}
\title{Cancel a fine-tuning job}
\usage{
fine_tune_cancel(job_id, provider)
}
\arguments{
\item{job_id}{A job ID from \code{\link[=fine_tune]{fine_tune()}}}

\item{provider}{The provider configuration}
}
\value{
The updated job object
}
\description{
Cancel a fine-tuning job
}
