% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{fine_tune_list}
\alias{fine_tune_list}
\title{List fine-tuning jobs}
\usage{
fine_tune_list(provider)
}
\arguments{
\item{provider}{The provider configuration}
}
\value{
A list of job objects (the API's \code{data} array)
}
\description{
List fine-tuning jobs
}
