% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{has_plugin}
\alias{has_plugin}
\title{Test whether a plugin is registered}
\usage{
has_plugin(kind, name)
}
\arguments{
\item{kind}{Plugin kind}

\item{name}{Plugin name}
}
\value{
\code{TRUE}/\code{FALSE}
}
\description{
Test whether a plugin is registered
}
