% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{job_cancel}
\alias{job_cancel}
\title{Cancel a running async job}
\usage{
job_cancel(job)
}
\arguments{
\item{job}{A job ID or \code{\link[=run_async]{run_async()}} result}
}
\value{
\code{NULL}, invisibly
}
\description{
Best-effort cancellation of an unresolved job (mirai may already be
executing the target).
}
