% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{job_status}
\alias{job_status}
\title{Report the status of an async job}
\usage{
job_status(job)
}
\arguments{
\item{job}{A job ID or \code{\link[=run_async]{run_async()}} result}
}
\value{
"running", "done", or "failed"
}
\description{
Report the status of an async job
}
