% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{list_plugins}
\alias{list_plugins}
\title{List registered plugins}
\usage{
list_plugins(kind = NULL)
}
\arguments{
\item{kind}{Optional kind to filter by}
}
\value{
A data.frame with columns \code{kind} and \code{name}
}
\description{
List registered plugins
}
