% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{merge_state}
\alias{merge_state}
\title{Reducer: merge state lists}
\usage{
merge_state(existing, new)
}
\arguments{
\item{existing}{Existing state list}

\item{new}{New values to merge}
}
\value{
Merged state
}
\description{
Reducer: merge state lists
}
