% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_stop}
\alias{metrics_stop}
\title{Stop a metrics server}
\usage{
metrics_stop(server)
}
\arguments{
\item{server}{A server handle from \code{\link[=start_metrics_server]{start_metrics_server()}}}
}
\value{
\code{NULL}, invisibly
}
\description{
Stop a metrics server
}
