% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{overwrite}
\alias{overwrite}
\title{Reducer: overwrite value}
\usage{
overwrite(existing, new)
}
\arguments{
\item{existing}{Existing value (ignored)}

\item{new}{New value}
}
\value{
The new value
}
\description{
Reducer: overwrite value
}
