% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{param_boolean}
\alias{param_boolean}
\title{Define a boolean parameter}
\usage{
param_boolean(description, required = TRUE)
}
\arguments{
\item{description}{Parameter description}

\item{required}{Whether the parameter is required}
}
\value{
A parameter definition
}
\description{
Define a boolean parameter
}
