% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{param_enum}
\alias{param_enum}
\title{Define an enum parameter}
\usage{
param_enum(description, values, required = TRUE)
}
\arguments{
\item{description}{Parameter description}

\item{values}{Allowed values}

\item{required}{Whether the parameter is required}
}
\value{
A parameter definition
}
\description{
Define an enum parameter
}
