% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{param_integer}
\alias{param_integer}
\title{Define an integer parameter}
\usage{
param_integer(description, required = TRUE)
}
\arguments{
\item{description}{Parameter description}

\item{required}{Whether the parameter is required}
}
\value{
A parameter definition
}
\description{
Define an integer parameter
}
