% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{param_number}
\alias{param_number}
\title{Define a number parameter}
\usage{
param_number(description, required = TRUE)
}
\arguments{
\item{description}{Parameter description}

\item{required}{Whether the parameter is required}
}
\value{
A parameter definition
}
\description{
Define a number parameter
}
