% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{param_object}
\alias{param_object}
\title{Define an object parameter}
\usage{
param_object(description, properties = list(), required = TRUE)
}
\arguments{
\item{description}{Parameter description}

\item{properties}{Named list of property definitions}

\item{required}{Whether the parameter is required}
}
\value{
A parameter definition
}
\description{
Define an object parameter
}
