% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{param_string}
\alias{param_string}
\title{Define a string parameter}
\usage{
param_string(description, required = TRUE)
}
\arguments{
\item{description}{Parameter description}

\item{required}{Whether the parameter is required}
}
\value{
A parameter definition
}
\description{
Define a string parameter
}
