% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2a.R
\name{print.agentgraph_a2a_server}
\alias{print.agentgraph_a2a_server}
\title{Print an A2A server handle}
\usage{
\method{print}{agentgraph_a2a_server}(x, ...)
}
\arguments{
\item{x}{An A2A server handle from \code{\link[=a2a_server]{a2a_server()}}}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print an A2A server handle
}
