% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{print.agentgraph_batch}
\alias{print.agentgraph_batch}
\title{Print a batch result}
\usage{
\method{print}{agentgraph_batch}(x, ...)
}
\arguments{
\item{x}{A batch result from \code{\link[=batch]{batch()}}}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a batch result
}
