% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{print.agentgraph_batch_job}
\alias{print.agentgraph_batch_job}
\title{Print a batch job}
\usage{
\method{print}{agentgraph_batch_job}(x, ...)
}
\arguments{
\item{x}{A job from \code{\link[=batch_submit]{batch_submit()}}}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a batch job
}
