% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{print.agentgraph_eval}
\alias{print.agentgraph_eval}
\title{Print an evaluation result}
\usage{
\method{print}{agentgraph_eval}(x, ...)
}
\arguments{
\item{x}{An evaluation result from \code{\link[=evaluate]{evaluate()}}}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print an evaluation result
}
