% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{print.agentgraph_job}
\alias{print.agentgraph_job}
\title{Print an async job}
\usage{
\method{print}{agentgraph_job}(x, ...)
}
\arguments{
\item{x}{A job from \code{\link[=run_async]{run_async()}}}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print an async job
}
