% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{print.agentgraph_prompt}
\alias{print.agentgraph_prompt}
\title{Print a prompt template}
\usage{
\method{print}{agentgraph_prompt}(x, ...)
}
\arguments{
\item{x}{A prompt template}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a prompt template
}
