% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{print.agentgraph_prompt_registry}
\alias{print.agentgraph_prompt_registry}
\title{Print a prompt registry}
\usage{
\method{print}{agentgraph_prompt_registry}(x, ...)
}
\arguments{
\item{x}{A prompt registry}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a prompt registry
}
