% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{print.agentgraph_serve}
\alias{print.agentgraph_serve}
\title{Print a REST server handle}
\usage{
\method{print}{agentgraph_serve}(x, ...)
}
\arguments{
\item{x}{A server handle from \code{\link[=serve_agent]{serve_agent()}}}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a REST server handle
}
