% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenant.R
\name{print.agentgraph_tenant}
\alias{print.agentgraph_tenant}
\title{Print a tenant configuration}
\usage{
\method{print}{agentgraph_tenant}(x, ...)
}
\arguments{
\item{x}{A tenant from \code{\link[=tenant]{tenant()}}}

\item{...}{Unused}
}
\value{
\code{x}, invisibly
}
\description{
Print a tenant configuration
}
