% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve_stop}
\alias{serve_stop}
\title{Stop a REST server}
\usage{
serve_stop(server)
}
\arguments{
\item{server}{A server handle from \code{\link[=serve_agent]{serve_agent()}}}
}
\value{
\code{NULL}, invisibly
}
\description{
Stop a REST server
}
