% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse.R
\name{sse_stop}
\alias{sse_stop}
\title{Stop an SSE server}
\usage{
sse_stop(server)
}
\arguments{
\item{server}{A server handle from \code{\link[=start_sse_server]{start_sse_server()}}}
}
\value{
\code{NULL}, invisibly
}
\description{
Stop an SSE server
}
