% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{state_graph}
\alias{state_graph}
\title{Create a new state graph}
\usage{
state_graph(entry, max_iterations = 25L)
}
\arguments{
\item{entry}{The entry point node ID}

\item{max_iterations}{Maximum number of iterations before stopping}
}
\value{
A graph object
}
\description{
Create a new state graph
}
