% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin.R
\name{unregister_plugin}
\alias{unregister_plugin}
\title{Unregister a plugin}
\usage{
unregister_plugin(kind, name)
}
\arguments{
\item{kind}{Plugin kind}

\item{name}{Plugin name}
}
\value{
\code{NULL}, invisibly
}
\description{
Unregister a plugin
}
