% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_store_clear}
\alias{vector_store_clear}
\title{Remove all vectors from a store}
\usage{
vector_store_clear(store)
}
\arguments{
\item{store}{A vector store}
}
\value{
The store (invisibly)
}
\description{
Remove all vectors from a store
}
