% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_store_count}
\alias{vector_store_count}
\title{Count vectors in a store}
\usage{
vector_store_count(store)
}
\arguments{
\item{store}{A vector store}
}
\value{
Number of stored vectors
}
\description{
Count vectors in a store
}
