% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_store_load}
\alias{vector_store_load}
\title{Load an hnswlib-backed store from disk}
\usage{
vector_store_load(store)
}
\arguments{
\item{store}{An hnswlib vector store (created with the same path/dim)}
}
\value{
The store (invisibly)
}
\description{
Load an hnswlib-backed store from disk
}
