% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_store_remove}
\alias{vector_store_remove}
\title{Remove a vector by ID}
\usage{
vector_store_remove(store, id)
}
\arguments{
\item{store}{A vector store}

\item{id}{The ID to remove}
}
\value{
The store (invisibly)
}
\description{
Remove a vector by ID
}
