% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_store_save}
\alias{vector_store_save}
\title{Persist an hnswlib-backed store to disk}
\usage{
vector_store_save(store)
}
\arguments{
\item{store}{An hnswlib vector store}
}
\value{
The store (invisibly)
}
\description{
Persist an hnswlib-backed store to disk
}
