% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{vector_store_search}
\alias{vector_store_search}
\title{Search a vector store for the k most similar vectors}
\usage{
vector_store_search(store, query, k = 5L)
}
\arguments{
\item{store}{A vector store from \code{create_vector_store()}}

\item{query}{Numeric query vector (the embedding)}

\item{k}{Number of results to return}
}
\value{
A data.frame with columns id, score, text, metadata
}
\description{
Search a vector store for the k most similar vectors
}
