% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-helpers.R
\name{visualize}
\alias{visualize}
\title{Visualize a graph structure}
\usage{
visualize(graph)
}
\arguments{
\item{graph}{A graph object}
}
\value{
A visualization (prints mermaid diagram)
}
\description{
Visualize a graph structure
}
